requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
pypdf
python-docx
numpy
httpx
openai
sentence-transformers
torch
pytest